Protect socket access using mutex
Access to socket read/write (particularily the ExecuteCommand, which
combines a socket write + read) must happen in an exclusive manner.

This adds a mutex around write/read of socket within ExecuteCommand.
Optionally, the caller can request to not to release the mutex in case
GetMessage must be called subsequently. Finally, the caller needs to
ensure to call ReleaseLock to release the mutex again.