Refactor GPIO Unit Tests
Refactor Pull up/down Tests.
1. Add one parameterized test (SetPullUpDown) which configures one GP
input (which is not connected to any driver) to either pull up or down.
2. Add SetPullUpDriver Test.
This tests implements an open drain logic by configuring a GPIO with
pull-up resistor (when acting as input) and switch the GPIO between
input and output to act as an open drain driver.