Make start and stop of GpioChangeListener explicit.
Since the GpioChangeListener requires a thread stopping the thread
requires explict shutdown on program exit.
This adds in addition to StartGpioChangeListener also the
StopGpioChangeListener which is used to stop the listener and the
associated thread. The thread is now longer started automatically during
class instantiation - this is now done in StartGpioChangeListener.

This also updates the WaitForSignalChange test case to use the new API
and adds a new test case ToggleCount which particularily tests the
RegisterLevelChangeCallback API.