Add UnregisterLevelChangeCallback method to Gpio class
This adds the UnregisterLevelChangeCallback method to Gpio class
to allow unregistering of callbacks also (so far only
RegisterLevelChangeCallback was existing)
RegisterLevelChangeCallback now returns an id (in fact hash code of the
internal GpioSubscriber class instance) which is used to unregister
the callback again.
Internally, this is achieved by replacing the existing list with a
Dictionary.

While changing this, also a mutex is added which protects all operations
on the Dictionary.