added an if-Statement to check if the desired Buffer size has been red on the Serial Interface
if the amount of red bytes is not the same as the desired amount an exception will be thrown.
this if statement has been introduced in order to debug an error which is occuring randomly
and not very often.
If the device has been resetted, sometimes the answer is only 2 Bytes long, but it should be
3 Bytes long.
The strange thing in this behaviour is, that the length byte has the correct length
stored but the receive buffer is not as long as the length byte specifies.