I2C: i2c_flags are sent as extended parameter via the socket
When opening the I2C bus, i2c_flags where sent along as integer parameter (p3) whereas
it should be sent as extended parameter. It was working up to now as the internal buffer within pigpiod was initialized with 0.

This fixes the issue by sending the i2c flags as extended parameter.