Add EventTimer class
This class allows running a timer which is started and stopped upon a certain configurable event:
For both, start and stop, two separate events (i.e. GPIO + level transition) can be configured.
Upon the start event (e.g. GPIO19 goes high) the timer is started and it is stopped once
the stop event occurs (e.g. GPIO 26 goes low).
Once the timer is stopped, the caller is informed using a callback, or, in case the Run method is used
this method is blocking and returns only when the timer is started and stopped.

Additionally, a max. waiting time can be specified to cover the case where timer is not started or not stopped.
In any case, the callback is executed (or Run() returns) and informs the user via the EventTimerResult object.